Demonstrate Ca-min x-inc burst delim doesn't work
There was a suggestion in the lab that using local minima in calcium
wherein the x variable is increasing should yield a method for detecting
when a burst has started or is ending for the purposes of delimitation.
From `toys/Delimitation failure.png` one can observe on the final spike
of a burst that this method should fail.